rename compat module
