+++ chanjo/compat.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-chanjo._compat
+chanjo.compat
 ~~~~~~~~~~~~~~~
 
 Python 2.7.x, 3.2+ compatability module.
